move test project, add a unit test
