--- UnitTests/UnitTests.csproj
@@ -1,40 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-
-    <IsPackable>false</IsPackable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <None Remove="Data\cycling_just_ride.json" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Content Include="Data\cycling_just_ride.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-      <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-    </Content>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="FluentAssertions" Version="6.7.0" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.3.1" />
-    <PackageReference Include="Moq" Version="4.18.2" />
-    <PackageReference Include="Moq.AutoMock" Version="3.4.0" />
-    <PackageReference Include="MSTest.TestFramework" Version="2.2.10" />
-    <PackageReference Include="NUnit" Version="3.13.3" />
-    <PackageReference Include="NUnit3TestAdapter" Version="4.2.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\src\Api\Api.csproj" />
-    <ProjectReference Include="..\src\Common\Common.csproj" />
-    <ProjectReference Include="..\src\Conversion\Conversion.csproj" />
-    <ProjectReference Include="..\src\Garmin\Garmin.csproj" />
-    <ProjectReference Include="..\src\PelotonToGarminConsole\PelotonToGarminConsole.csproj" />
-    <ProjectReference Include="..\src\Peloton\Peloton.csproj" />
-  </ItemGroup>
-
-</Project>
