+++ src/UnitTests/Common/Helpers/WorkoutHelperTests.cs
