+++ src/UnitTests/Data/github_responses/latest_release_response.json
@@ -1,0 +1,74 @@
+﻿{
+  "url": "https://api.github.com/repos/philosowaffle/peloton-to-garmin/releases/75711223",
+  "assets_url": "https://api.github.com/repos/philosowaffle/peloton-to-garmin/releases/75711223/assets",
+  "upload_url": "https://uploads.github.com/repos/philosowaffle/peloton-to-garmin/releases/75711223/assets{?name,label}",
+  "html_url": "https://github.com/philosowaffle/peloton-to-garmin/releases/tag/v3.0.3",
+  "id": 75711223,
+  "author": {
+    "login": "github-actions[bot]",
+    "id": 41898282,
+    "node_id": "MDM6Qm90NDE4OTgyODI=",
+    "avatar_url": "https://avatars.githubusercontent.com/in/15368?v=4",
+    "gravatar_id": "",
+    "url": "https://api.github.com/users/github-actions%5Bbot%5D",
+    "html_url": "https://github.com/apps/github-actions",
+    "followers_url": "https://api.github.com/users/github-actions%5Bbot%5D/followers",
+    "following_url": "https://api.github.com/users/github-actions%5Bbot%5D/following{/other_user}",
+    "gists_url": "https://api.github.com/users/github-actions%5Bbot%5D/gists{/gist_id}",
+    "starred_url": "https://api.github.com/users/github-actions%5Bbot%5D/starred{/owner}{/repo}",
+    "subscriptions_url": "https://api.github.com/users/github-actions%5Bbot%5D/subscriptions",
+    "organizations_url": "https://api.github.com/users/github-actions%5Bbot%5D/orgs",
+    "repos_url": "https://api.github.com/users/github-actions%5Bbot%5D/repos",
+    "events_url": "https://api.github.com/users/github-actions%5Bbot%5D/events{/privacy}",
+    "received_events_url": "https://api.github.com/users/github-actions%5Bbot%5D/received_events",
+    "type": "Bot",
+    "site_admin": false
+  },
+  "node_id": "RE_kwDOBuLlFs4Eg0L3",
+  "tag_name": "v3.0.3",
+  "target_commitish": "master",
+  "name": "v3.0.3",
+  "draft": false,
+  "prerelease": false,
+  "created_at": "2022-08-30T01:23:31Z",
+  "published_at": "2022-08-30T01:38:24Z",
+  "assets": [
+    {
+      "url": "https://api.github.com/repos/philosowaffle/peloton-to-garmin/releases/assets/76252280",
+      "id": 76252280,
+      "node_id": "RA_kwDOBuLlFs4Ei4R4",
+      "name": "win10-x64_3.0.3.zip",
+      "label": "",
+      "uploader": {
+        "login": "github-actions[bot]",
+        "id": 41898282,
+        "node_id": "MDM6Qm90NDE4OTgyODI=",
+        "avatar_url": "https://avatars.githubusercontent.com/in/15368?v=4",
+        "gravatar_id": "",
+        "url": "https://api.github.com/users/github-actions%5Bbot%5D",
+        "html_url": "https://github.com/apps/github-actions",
+        "followers_url": "https://api.github.com/users/github-actions%5Bbot%5D/followers",
+        "following_url": "https://api.github.com/users/github-actions%5Bbot%5D/following{/other_user}",
+        "gists_url": "https://api.github.com/users/github-actions%5Bbot%5D/gists{/gist_id}",
+        "starred_url": "https://api.github.com/users/github-actions%5Bbot%5D/starred{/owner}{/repo}",
+        "subscriptions_url": "https://api.github.com/users/github-actions%5Bbot%5D/subscriptions",
+        "organizations_url": "https://api.github.com/users/github-actions%5Bbot%5D/orgs",
+        "repos_url": "https://api.github.com/users/github-actions%5Bbot%5D/repos",
+        "events_url": "https://api.github.com/users/github-actions%5Bbot%5D/events{/privacy}",
+        "received_events_url": "https://api.github.com/users/github-actions%5Bbot%5D/received_events",
+        "type": "Bot",
+        "site_admin": false
+      },
+      "content_type": "raw",
+      "state": "uploaded",
+      "size": 51874281,
+      "download_count": 108,
+      "created_at": "2022-08-30T01:38:24Z",
+      "updated_at": "2022-08-30T01:38:26Z",
+      "browser_download_url": "https://github.com/philosowaffle/peloton-to-garmin/releases/download/v3.0.3/win10-x64_3.0.3.zip"
+    }
+  ],
+  "tarball_url": "https://api.github.com/repos/philosowaffle/peloton-to-garmin/tarball/v3.0.3",
+  "zipball_url": "https://api.github.com/repos/philosowaffle/peloton-to-garmin/zipball/v3.0.3",
+  "body": "## Features\r\n\r\n- [#320] Add support for Distance based cycling workouts\r\n\r\n## Fixes\r\n\r\n- [#305] Fix logging and error handling when a Deserialization error occurrs\r\n- [#304] Fix Deserialization error when FTP source is `ftp_estimated_source`\r\n- [#309] Fix `Found unknown distance unit m` and `kph`\r\n- [#323] Fix scenario where Fire TV workout had null End Time and failed to deserialize\r\n\r\n## Changes\r\n\r\n- [#311] Sanitize Verbose logging - Prometheus HTTP labels also modified to better strip dynamic values"
+}
