+++ src/UnitTests/GitHub/GitHubServiceTests.cs
@@ -1,0 +1,17 @@
+﻿using NUnit.Framework;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace UnitTests.GitHub;
+
+public class GitHubServiceTests
+{
+	[Test]
+	public async Task GetLatestReleaseAsync()
+	{
+		Assert.Fail();
+	}
+}
