+++ src/UnitTests/Peloton/PelotonServiceTests.cs
